Moved master component to components folder Added structural directive *ngIf to load component based on condition.
