+++ angular_18_tutorial/src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { RolesComponent } from './components/roles/roles.component';
-import { MasterComponent } from './master/master.component';
+import { MasterComponent } from './components/master/master.component';
 
 @Component({
   selector: 'app-root',
