--- angular_18_tutorial/src/app/master/master.component.html
@@ -1,14 +1,0 @@
-<div class="row">
-    <div class="col-12 text-center">
-        <div class="btn-group">
-            <button type="button" class="btn btn-primary">Roles</button>
-            <button type="button" class="btn btn-primary">Designation</button>
-        </div>
-    </div>
-</div>
-<div class="row">
-    <div class="col-12" *ngIf="currentComponent == 'Roles'">
-        <app-roles></app-roles>
-
-    </div>
-</div>
